requirements.txt
absl-py==1.4.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
Django==4.2.2
django-jazzmin==2.6.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
html5lib==1.1
idna==3.4
jax==0.4.13
keras==2.12.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
playsound==1.3.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.11.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
xhtml2pdf==0.2.11
